Fix build time macro use related to phasing...
... there is some evaluation that I expected to happen at compile
time, but only happens at execution time.

The bug happens at execution time trying to do (substring something 0
-1) that happens when the branch does not have a tag?

In anycase, part of the evaluation is done at compile time
e.g. retrieving the current date with the macro `include-date`, the
variable `build-date` is only computed at execution time. I tried to
make build-date a lambda, same for the related tag variable, and it
failed in a way that is more scary.