letloop: add letloop root...
... add letloop root as part of the cli. Only emulate did not work
during my tests due to a problem with a lazy kernel? Anyway, it should
be enough for what is next, see #6.

This commit include:

- Drop -exec suffix in procedure names;
- Add print to root-available;
- Fix system* error symbol, and include variables;
- sudo must be explicit;

The bulk of the cli massage is done in the new procedure called
letloop-root. More massage is warranted, for instance to support
setting environment with KEY=VALUE, but not today.